bug fixed - centroids to doubles

+++ wavfind.cpp
@@ -40,9 +40,9 @@ int main(int argc, char *argv[])
 	}
 
 	string codebooksPath{argv[argc - 1]};
-	vector<vector<vector<short>>> codebooks;
+	vector<vector<vector<double>>> codebooks;
 	vector<tuple<string, string>> filesNames;
-	vector<short> blocksSizes;
+	vector<int> blocksSizes;
 
 	DIR *dir;
 	struct dirent *ent;
@@ -89,17 +89,17 @@ int main(int argc, char *argv[])
 								cerr << "Invalid codebook file structure" << endl;
 								return 1;
 							}
-							codebooks.push_back(vector<vector<short>>(codebookSize));
+							codebooks.push_back(vector<vector<double>>(codebookSize));
 							lineIdx++;
 							continue;
 						}
 
-						codebooks[fileIdx][lineIdx] = vector<short>(blocksSizes[fileIdx]);
+						codebooks[fileIdx][lineIdx] = vector<double>(blocksSizes[fileIdx]);
 
 						while ((pos = line.find(delimiter)) != string::npos)
 						{
 							value = line.substr(0, pos);
-							codebooks[fileIdx][lineIdx][valueIdx] = stoi(value);
+							codebooks[fileIdx][lineIdx][valueIdx] = stod(value);
 							line.erase(0, pos + delimiter.length());
 							valueIdx++;
 						}
